Add file for the prompt's section code
